requirements.txt
# Core dependencies
streamlit==1.31.1
langchain==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-groq==0.2.0
langgraph==0.2.28
python-dotenv==1.0.1

# Data & Visualization
pandas==2.2.0
plotly==5.19.0
matplotlib==3.8.3
seaborn==0.13.2

# Database
# sqlite3 is built-in with Python, no need to install

# Utilities
typing-extensions==4.9.0
pydantic==2.6.1

# Testing (development)
pytest==8.0.0
pytest-cov==4.1.0
pylint==3.0.3
